backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
sqlite3
tensorflow==2.13.0
opencv-python==4.8.0.74
numpy==1.24.3
Pillow==10.0.0
scikit-learn==1.3.0
xgboost==1.7.6
pandas==2.0.3
joblib==1.3.2